backend/requirements.txt
fastapi
uvicorn
pandas
joblib
scikit-learn
python-dotenv
python-multipart
openpyxl
